day 11
phew this day was hard... to do in Rust!

had a hard time banging my head trying to do cross-references of items, and having problems with the borrow checker.

Hopefully a graph library helped me simplifying the process (even if not as simple as in Python.)

I even had a x10 perfomance improvement against my initial naive implementation :)