day 15
As my p1 solution was so slow for p2, I was too focused finding an optimal solution that wouldn't loop 30000000 times.
So just searching for a hint expecting a complex modular arithmetic property like previous exercises, I spoiled myself just by seeing a reddit user saying "I just used hashmap".
So the answer to improve performance was obvious and I felt dumb not implementing it earlier.

Apparently this is a Van Eck sequence, that can't be simplified any way.

Still on my computer the answer would take 20s to compute.
That's where I discovered that the "--release" option of cargo is not just for show, and that it optimizes performances 10 times.

Put some tests as "ignored", because each file still takes  2-3s to solve, and I have a pre-commit hook with the tests... I'm bored waiting for tests guaranteed to work.