day 17
not really optimized solution, forgot tu use Hashmaps instead of vectors.
At least I noticed the pattern was parallel on the z and w axis.

Got a bit tired when I saw the twist of part 2.
My solution, instead of making it generic, was to add 1 more "Vec" and make part 1 a special case with only1 level of "w" dimension.
I feel it's not beautiful at all, but it does the job.

Don't ask me to make a 5-dimensions puzzle...

maybe I'lkl make a 3D render of it if i feel the force to do so, but I may make a pause in my AoC exercices for a few days.