3d render of day 17 part 1
took way too much time to do it, but it's done.

use Kiss3d which is simple to use but lacks some features to be honest. (forced to use a big texture of a checker instead of repeating the texture, and only one light not very beautiful)

Only 1 animation because vertice change animations apply to all cubes, without distinctions.